+++ changelog.md
@@ -12,8 +12,10 @@ Codion Change Log
 ### is.codion.framework.servlet
 - EntityService.FunctionHandler and ProcedureHandler.json() added.
 - EntityService.deserialize() removed, Serializer.deserialize() used instead.
+- EntityService.JSON (default true) and SERIALIZATION (default false) configuration properties added, other properties renamed.
 ### is.codion.framework.db.http
 - JsonHttpEntityConnection, JSON based function and procedure execution added.
+- HttpEntityConnection.JSON default back to being true.
 
 ## 0.18.56
 ### is.codion.common.core